+++ examples/resources/sdwan_custom_application/resource.tf
@@ -1,5 +1,5 @@
 resource "sdwan_custom_application" "example" {
-  app_name     = "Example"
+  app_name     = "Example-Custom"
   server_names = ["*customapp.com"]
   l3l4 = [
     {