+++ examples/resources/sdwan_service_lan_vpn_interface_ethernet_feature/resource.tf
@@ -17,6 +17,7 @@ resource "sdwan_service_lan_vpn_interface_ethernet_feature" "example" {
   ]
   ipv4_dhcp_helper  = ["1.2.3.4"]
   ipv6_address_type = "static"
+  ipv6_address      = "2001:0:0:1::0/64"
   ipv6_dhcp_helpers = [
     {
       address           = "2001:0:0:1::0"